compartmentalized navigators
